feat: always return the instance from init()
NOTE: There is a divergence from the book. As the methods `getAt` and
`assignAt` takes a Token instead of a string as the name (in order to
keep using RuntimeError), we cannot easily fetch the class instance
easily by searching for a variable named "this". Instead, we store the
instance as a property in LoxFunction in order to fetch it directly when
needed (e.g. returning it when invoking init()).